new: full system for export

+++ libft/ft_strncmp.c
@@ -6,15 +6,15 @@
 /*   By: algasnie <algasnie@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/15 13:53:13 by algasnie          #+#    #+#             */
-/*   Updated: 2025/10/22 18:17:19 by algasnie         ###   ########.fr       */
+/*   Updated: 2026/03/03 12:07:11 by algasnie         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strncmp(const char *s1, const char *s2, size_t n)
+int	ft_strncmp(const char *s1, const char *s2, int n)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	if (n == 0)
